requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
psycopg[binary]
python-jose[cryptography]>=3.3.0
google-auth>=2.30.0
requests>=2.32.0
pydantic>=2.9.0
python-dotenv